Rescue exceptions in dbus_methods and reply with DBus errors instead of crashing.
The attached patch forwards Exceptions that occur in DBus service calls to the
sender, by replying with a org.freedesktop.DBus.Error.Failed message instead
of making the service crash.

I am not absolutely convinced that this is the right way to handle service
sind exceptions, but it's the only one i could find. In any case, a way to pass
error conditions to the called must be implemented. This patch is my
suggestion how to do it.

[whitespace adjusted]