WIP trying a new main loop
now the tests that do have conections in the loop fail,
because wait_for_message is private and unfinished. Called by send_sync :Hello already.
But if we replaced it qith cq.pop, it would deadlock because cq is only
filled later in the main loop. it seems the main loop cannot be the only
filling place.