Ensure DBus.type constructs a valid Type.
Previously it was easy to construct an invalid signature,
even for trivially bad cases.

Now, make sure `DBus.type` constructs a valid Single Complete Type,
and add `DBus.types` for constructing a valid Signature (0 or more SCTs).

Add test cases.

The corresponding section of the specification is
https://dbus.freedesktop.org/doc/dbus-specification.html#type-system

Also, `DBus::Type` was an internal Module and the constructed Class
was awkwardly named `DBus::Type::Type`.
Merged them (with a compatibility "symlink")