Service side emits_changed_signal, affecting PropertiesChanged
At the service side, allow declaring that `PropertiesChanged` will send
only the property name (`:invalidates`) or that the signal will not be
emitted at all, either because the property is a constant (`:const`) or
for another reason (`false`). If `true` it sends the signal with the
property value.

See also the description of the
`org.freedesktop.DBus.Property.EmitsChangedSignal` annotation at
<https://dbus.freedesktop.org/doc/dbus-specification.html#introspection-format>

Again remove the type argument from dbus_watcher and the
dbus_property_changed4 method, the optimization is not worth the API
complication.