DBus::Object#object_server replaces @service
(which still works)
and the short-lived @connection

This matters when implementing (and testing) Objects that manage other Objects.

Also raise a better exception than "NoMethodError on nil" when emitting a
signal on an unexported object, for example by assigning its property.
(Happened in Agama tests)

Speaking about unexported objects:

```
obj = DBus::Object.new("/path")
object_server.export(obj)
```

lets you construct objects that do have a path but are not exported at it, so
in retrospect a better API would be `export(DBus::Object.new, "/path")` but
it's too late and not a big deal.